requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binarypsycopg2-binary
python-dotenv
cloudinary
torch
torchvision
Pillow